feat(story): add story for downloading data
